Create Mercari Price Challenge
This challenge was posted on Kaggle platform on 23rd November, 2017 for $100,000 to predict the selling price of a listing based on information a user provides for that listing. The dataset has close to 750k rows with 8 features in Stage 1. For Stage 2, the code will be implemented on over 2.5 million rows. As a single contributor I have tried to compete in this challenge.